I modified userspace to work with python because why not?!
So far the only testing I've done is to diff the output of userspace.py and userspace-py3.py a few times.  There are 5 items who's values change and they are things that look like the value should be changing.  I will push a changelog of what I changed and which 5 things change values.